Auto-hide tip amount and total amount
Hide the tip amount and total amount if they do not accurately reflect
the current bill amount and tip percentage in the UI, or reveal them if
they are accurate. With of the "tip button" design of this app, it is a
nice feature to hide inaccurate amounts until the tip button is pressed.